🎨 extract checkForHalts from automation index into separate file
